Update libvpx and incorporate new build system
Move libvpx down a directory. Separate libwebm and rename it. It would
be more accurate to rename external/libvpx to external/webm.

Use file lists directly from upstream libvpx. This allows adding new
targets and new features (such as the encoder) easily. MIPS and x86
options are in progress. See new file "UPDATING"

The new libvpx checkout is not from a release branch. The decoder is
stable but it should be checked and potentially updated if the encoder
is enabled.

Requires I42b51e2845a696a6e211dde00951afc8f571336f which updates
libstagefright to account for new paths and library names.

Change-Id: I739f99d48b8d7e6354c416ef2ca79c954826307f