Update libvpx and incorporate new build system
Move libvpx down a directory. Separate libwebm and rename it. It would
be more accurate to rename external/libvpx to external/webm.

Use file lists directly from upstream libvpx. This allows adding new
targets and new features (such as the encoder) easily. MIPS and x86
options are in progress. See new file "UPDATING"

The new libvpx checkout is not from a release branch. The decoder is
stable but it should be checked and potentially updated if the encoder
is enabled.

Requires Iec8566bce0a4763a9a17b4a46f15aa39b5bcd141 which updates
libstagefright to account for new paths and library names.

Change-Id: I2c0a87b0fc049e855e21f455b806287b578091de