Add -no-integrated-as at local level.
Later we will enable integrated-as as default at the global level.

Also pass include path explicitly to the GCC assembler.
Unlike gcc, clang doesn't pass -I flags on to the assembler - causing
the .include statements not to find the generated asm files.

This is reported as a bug in clang at
http://llvm.org/bugs/show_bug.cgi?id=21000

BUG: 17820427
BUG: 17137906
Change-Id: I788f3c8a3ca71948476c469b21dfda14483d269d